% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_net_profit.R
\name{compute_profit}
\alias{compute_profit}
\title{compute net profit'}
\usage{
compute_profit(T, P, irr, discount, price, cost, crop.pars)
}
\arguments{
\item{T}{annual temperature}

\item{P}{annual precip}

\item{irr}{}

\item{discount}{}

\item{price}{of fruit}

\item{cost}{of irrigation}
}
\value{
profit
}
\description{
Function to compute profits
}

