% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{contributiontopeak}
\alias{contributiontopeak}
\title{Calculates the contribution of each cell to peak flow}
\usage{
contributiontopeak(ft, basins, bysize = TRUE)
}
\arguments{
\item{ft}{a raster object, two-dimensional array or matrix of flow times as returned by \code{\link[=flowtimes]{flowtimes()}}}

\item{basins}{a raster object, two-dimensional array or matrix with basins numbered as integers as
returned by \code{\link[=basindelin]{basindelin()}}.}

\item{bysize}{an optional logical indicating whether or not to weight values by basin size}
}
\value{
a raster object of the contribution of each cell to peak flow
}
\description{
\code{contributiontopeak} calculates the contribution of each cell of a digital elevation dataset to
peak flow
}
\details{
contribution to peak flow is calculated by fitting a log-normal distribution
to flow times, estimating the probability density function (pdf) of this distribution, weighted
by the maximum value of of the pdf. If \code{bysize} is TRUE, output values are multiplied
by the size of the basin thereby making the assumption that larger basins have higher
peak flow. Alternatively real peak flows can be estimated from meterological
data.
}
\examples{
basins <- basindelin(dtm100m) # Takes ~20 seconds to run
ft <- flowtimes(dtm100m)
plot(contributiontopeak(ft, basins))
plot(contributiontopeak(ft, basins, bysize = F))

}
