% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornwallrain}
\alias{cornwallrain}
\title{A 1 km resolution raster object of rainfall in west Cornwall, UK in January 2015.}
\format{A raster object with 77 rows and 116 columns.}
\source{
\url{https://eip.ceh.ac.uk/rainfall/}
}
\usage{
cornwallrain
}
\description{
A raster object containing rainfall (mm) with sea coded as NA for the area bounded by
82500, 198500, 4500, 81500  (xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid
Reference system (CRS: +init=epsg:27700).
}
\keyword{datasets}
