% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{covarMBLRPM}
\alias{covarMBLRPM}
\title{Internal funcion used to calculate covariance of rainfall}
\usage{
covarMBLRPM(a, l, v, k, f, mx, h = 1, lag = 1)
}
\description{
Internal funcion used to calculate covariance of rainfall
}
