% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{cre.daily}
\alias{cre.daily}
\title{Calculates crop reference evapotranspiration from daily data}
\usage{
cre.daily(Rn, tmin, tmax, u1, rhmin, rhmax, p)
}
\arguments{
\item{Rn}{a single numeric value, vector, two-dimensional array or matrix of net radiation (MJ / m^2 / day)}

\item{tmin}{a single numeric value, vector, two-dimensional array or matrix of daily minimum temperature (deg C)}

\item{tmax}{a single numeric value, vector, two-dimensional array or matrix of daily maximum temperature (deg C)}

\item{u1}{a single numeric value, vector, two-dimensional array or matrix of wind speed at one metre above the ground (m /s)}

\item{rhmin}{a  single numeric value, vector, two-dimensional array or matrix of daily minimum relative humidity (percentage)}

\item{rhmax}{a  single numeric value, vector, two-dimensional array or matrix of daily maximum relative humidity (percentage)}

\item{p}{a single numeric value, vector, two-dimensional array or matrix of pressure (Pa)}
}
\value{
a single numeric value, vector, two-dimensional array or matrix of crop reference evapotranspiration (mm / day)
}
\description{
\code{cre.daily} applies the Penman-Monteith equation to derive hourly crop reference evapotranspiration
}
\details{
Applies the method detailed in Allen et al (1998). If relatively fine-resolution spatial data are needed,
radiation, temperature, wind and relative humidity can be downscaled using the microclima package. Humidity conversions
can also be performed using this package. If only sea-level pressure is known, then
\code{p} can be derived using function \code{\link[=pressure.height]{pressure.height()}}. To convert radiation from watts m^-2 to MJ m^-2 day^-1 multiply
by 0.0864.
}
\examples{
cre.daily(13.2, 10, 20, 1.3, 60, 100, 101300)
}
\seealso{
cre.hourly for deriving crop reference evapotranspiration from hourly data
}
