% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{cre.hourly}
\alias{cre.hourly}
\title{Calculates crop reference evapotranspiration from hourly data}
\usage{
cre.hourly(Rn, tc, u1, rh, p, dn)
}
\arguments{
\item{Rn}{a single numeric value, vector, two-dimensional array or matrix of net radiation (MJ / m^2 / hr)}

\item{tc}{a single numeric value, vector, two-dimensional array or matrix of temperature (deg C)}

\item{u1}{a single numeric value, vector, two-dimensional array or matrix of wind speed at one metre above the ground (m /s)}

\item{rh}{a single numeric value, vector, two-dimensional array or matrix of relative humidity (percentage)}

\item{p}{a single numeric value, vector, two-dimensional array or matrix of pressure (Pa)}

\item{dn}{binary variable indicating whether day or night (day - 1, night = 0). Must be
a single numeric value or have the same dimensions as Rn}

\item{u1}{a single numeric value, vector, two-dimensional array or matrix of wind speed at one metre above the ground (m /s)}
}
\value{
a single numeric value, vector, two-dimensional array or matrix of crop reference evapotranspiration (mm / hr)
}
\description{
\code{cre.hourly} applies the Penman-Monteith equation to derive hourly crop reference evapotranspiration
}
\details{
Applies the method detailed in Allen et al (1998). If relatively fine-resolution spatial data are needed,
radiation, temperature, wind and relative humidity can be downscaled using the microclima package. The variable \code{dn}
and humidity conversions can also be obtained using this package. If only sea-level pressure is known, then
\code{p} can be derived using function \code{\link[=pressure.height]{pressure.height()}}. To convert radiation from watts m^-2 to MJ m^-2 hr^-1 multiply
by 0.0036.
}
\examples{
cre.hourly(1.1, 20, 1.3, 70, 101300, 1) # day
cre.hourly(-0.3, 5, 1.3, 100, 101300, 0) # night (cold and wet)
cre.hourly(-0.1, 30, 1.3, 60, 101300, 0) # night (hot and dry)
}
\seealso{
cre.daily for deriving crop reference evapotranspiration from daily data
}
