% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dailyrain}
\alias{dailyrain}
\title{Daily precipitation}
\format{A data.frame with 12053 rows and two variables.
\describe{
\item{obs_time}{Date of observation}
\item{precipitation}{Daily precipitation total (mm)}
}}
\source{
\url{https://eip.ceh.ac.uk/rainfall/}
}
\usage{
dailyrain
}
\description{
A dataset containing daily precipitation on the Lizard Peninsula for the period 1982-2015
}
\keyword{datasets}
