% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{.basinchars}
\alias{.basinchars}
\title{Internal function for calculating basin characteristics}
\usage{
.basinchars(md, mb, sea = F)
}
\description{
Internal function for calculating basin characteristics
}
