% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydromodel.R
\name{Dtheta}
\alias{Dtheta}
\title{Hydraulic diffusivity}
\usage{
Dtheta(sm, sp, timestep = 86400)
}
\arguments{
\item{sm}{Soil water fraction}

\item{sp}{a list of van Genuchten soil parameters as returned by \code{\link[=getsoilparams]{getsoilparams()}}}

\item{timestep}{number of seconds in each time step of model run (default one day)}
}
\value{
Hydraulic conductivity (cm / time-step)
}
\description{
\code{Dtheta} calculates hydrualic diffusivity
}
\details{
Applies the van Genuchten equation to estimate the hydrualic diffusivity
(van Genuchten 1980 Soil Science Society of America Journal 44:892-898).
Paramaters of the van Genuchten equation can be found in the dataset \code{\link[=soilparams]{soilparams()}}.
Ksat and the returned value are adjusted so that returned units are in cm per \code{timestep}.
\code{Dtheta} is contrained to be finite by ensuring that \code{sm} marginally less than \code{Smax}
when saturated and is assumed to be zero, when \code{sm} = \code{Smin}.
}
\examples{
sm <- c(74:422) / 1000
plot(log(Dtheta(sm, getsoilparams("Loam")), 10) ~ sm)

}
\seealso{
\code{\link[=soilparams]{soilparams()}}
}
