\name{ecohydrotools-package}
\alias{ecohydrotools-package}
\alias{ecohydrotools}
\docType{package}
\title{
\packageTitle{ecohydrotools}
}
\description{
\packageDescription{ecohydrotools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ecohydrotools}
\packageIndices{ecohydrotools}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{ecohydrotools}

Maintainer: \packageMaintainer{ecohydrotools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
