% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{findBLpar}
\alias{findBLpar}
\title{Estimate Bartlett-Lewis model parameters for deriving sub-daily rainfall}
\usage{
findBLpar(rainseq, period = "daily")
}
\arguments{
\item{period}{a desciption of the time-interval between successive values of \code{rainseq}.
One of \code{hourly}, \code{sixhourly} or \code{daily}. Other time intervals are not supported.}

\item{raindseq}{a vector time-series of rainfall}
}
\value{
A data.frame with 10 elements. The first six are the Bartlett-Lewis model parameters
(see details). Additionally, \code{mean.rain}: the modelled mean intensity of rainfall in \code{rainseq}
as derived using the parameters of Bartlett-Lewis model; \code{var.rain}: the modelled variance in
rainfall;  \code{cvar.rain}, the modelled covariance between rainfall and rainfall lagged by one
time-step; and  \code{prop.dry}, the modelled proportion of elements in \code{rainseq} that have no rain. These are not used for generating sub-daily rainfall, but are used to check that parameter estimates are sensible (see details).
}
\description{
\code{findBLpar} is used to estimate Bartlett-Lewis model parameters needed by function \code{\link[=subdailyrain]{subdailyrain()}} to derive sub-daily rainfall.
}
\details{
The function is based on the Bartlett-Lewis Rectangular Pulse model described by
Rodriguez-Iturbe (1987 & 1988). The model has six parameters and is characterized as a
particular form of clustering process in which each cluster of rainfall events (hereafter storms)
consists of one or more rainfall cells being generated in the start of the process. The
parameter \code{l} governs the frequency of storms; \code{k} is used to control to start of rainfall
events associated with each storms and \code{f} is used to is used to control the time at which
rainfall ceases; \code{mx} represents intensity of rainfall associated with storms and \code{a} and \code{v}
are scale and shape parameters used to control variation in the duration of storms.

Once know, the parameters can be used to calculate, for any time-interval, the mean and variance
in the intensity of rain, the autocorrelation rainfall intensities for any lag period, and the
proportion of dry days or hours. Since these vary seasonally, or by month, normally the model
parameters are derived separately for each month using time-series of data covering multiple
years (see example 2).

These variables can also be calculated analytically from the sequence of rainfall data, but at
only at time intervals ≥ that of the data. This function derives the parameters through
iteration. In each iteration, 1000 parameters are generated randomly from uniform distributions
within finite limits and the mean, variance, covariance and dry proportion estimated and
compared to those obtained analytically using an objective function. This comparison is used to
narrow the limits from which parameters can be generated and algorithm iterated until
convergence.

It remains challenging to generate six parameters using a limited amount of information against
which to ensure a unique optimum of the objective, and hence convergence of during optimization.
Initial parameter limits of chosen carefully to assist with optimisation, but the function is
crude, not particularly computationally efficient, and not widely tested. It does, however, work
in most circumstances, and the estimated mean, variance, covariance and dry proportion of rainfall
events are also given to enable sensible cross-checking. Some element of the coding from the
HyetosMinute package, available here under a GNU General Public License:
http://www.itia.ntua.gr/en/softinfo/3/, have been used.
}
\examples{
# ========= Example 1: ============ #
# Find Bartlett-Lewis model parameters for March rainfall
tme <- as.POSIXlt(dailyrain$obs_time)
marchrain <- dailyrain$precipitation[which(tme$mon + 1 == 3)]
findBLpar(marchrain) # Takes ~ 30 seconds
# ========= Example 2: ============ #
# Find Bartlett-Lewis model parameters for each month
# Warning: Takes a few minutes to run
tme <- as.POSIXlt(dailyrain$obs_time)
pars <- list()
for (mon in 1:12) {
  monthrain <- dailyrain$precipitation[which((tme$mon + 1) == mon)]
  pars[[mon]] <- findBLpar(marchrain)
 print(pars[[mon]])
}

}
\references{
Rodriguez-Iturbe I, Cox DR & Isham V (1987) Some models for rainfall based on stochastic point
processes. Proc. R. Soc. Lond., A 410: 269-288.

Rodriguez-Iturbe I, Cox DR & Isham V (1988) A point process model for rainfall: Further
developments. Proc. R. Soc. Lond., A 417: 283-298.
}
