% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{findm}
\alias{findm}
\title{Finds coefficient for scaling topographic wetness to flow velocity}
\usage{
findm(tpx)
}
\arguments{
\item{tpx}{a raster object, two-dimensional array or matrix of topographic wetness
as returned by \code{\link[=topidx]{topidx()}}}
}
\value{
a numeric scaling ceofficient
}
\description{
\code{findm} is an internal function used by \code{\link[=flowvelocity]{flowvelocity()}} to scale topographic
wetness to flow velocity.
}
\examples{
findm(topidx(dtm100m))

}
