% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{flowdir}
\alias{flowdir}
\title{Calculates flow direction}
\usage{
flowdir(dem)
}
\arguments{
\item{dem}{a raster object, two-dimensional array or matrix of elevations.}
}
\value{
a raster object, two-dimensional array or matrix of flow directions (1-9).
}
\description{
\code{flowdir} is used to calculate the direction of flow from every cell of a digital
elevation dataset
}
\details{
Flow direction is expressed as given by \code{array(c(1:9), dim = c(3,3))}. I.e.
1 = NW, 2 = W, 3 = SW, 4 = N, 5 = to self, 6 = S, 7 = NE, 8 = E and 9 = SE. Flow direction is determined
by which adjacent cell has the lowest elevation.
}
\examples{
dem <- aggregate(dtm1m, 20)
plot(flowdir(dem), main = "Flow direction")

}
