% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{flowtimes}
\alias{flowtimes}
\title{Calculates flow time to basin bottom}
\usage{
flowtimes(dem, nr = 0.025, wf = 5e-04, maxiter = 100,
  minslope = atan(0.005/mean(res(dem))))
}
\arguments{
\item{dem}{a raster object of elevations}

\item{nr}{an optional single numeric value, or a raster object, two-dimensional array or matrix
of Manning's roughness coefficients (see details)}

\item{wf}{n optional single numeric value, or a raster object, two-dimensional array or matrix of
mean water volumes (see details)}

\item{maxiter}{a positive integer specifying the maximum number of iterations allowed when flow
path algorithm gets stuck in a loop.}

\item{minslope}{an optional positive value specifying the minimum slope allowed (see details).}
}
\value{
a raster object of flow times (s)
}
\description{
\code{flowtimes} calculates flow paths and applies the flow velocity function to calculate the
the time taken for water from each cell to reach the lowest point in the basin. It is used
to calculate which cells contribute to peak flow.
}
\details{
If \code{wf} is a single value it should approximate the mean proportion of land wetted.
If \code{wf} is an array, matrix or raster object, then values should approximate the mean proportion of
land wetted seperately for each basin, but should not be variable within basins.
If slope is zero, infinite topographic wetness index values are returned. Thus,
flat areas are assigned a value specifed by \code{minslope}. The default value, given by
atan(0.005 / mean(res(dem))), assumes that height differences between adjacent cells exceed
0.005. Accumulated flow is multiplied by the cell dimensions, thus\code{dem} should have an
equal area projection. Values of \code{nr} fora variaty of substrate types can be found
here: \code{\link{http://www.fsl.orst.edu/geowater/FX3/help/8_Hydraulic_Reference/Mannings_n_Tables.htm}}.
}
\examples{
plot(log(flowtimes(dtm100m)), main = "Log flow time (s)")

}
