% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydromodel.R
\name{getsoilparams}
\alias{getsoilparams}
\title{Gets van Genuchten soil paramaters}
\usage{
getsoilparams(soiltype)
}
\arguments{
\item{soiltype}{a single character, matrix or raster of characters as listed in the dataset \code{\link[=soilparams]{soilparams()}}}
}
\value{
a list of van Genuchten parameters (see details)
}
\description{
\code{getsoilparams} Obtains van Genuchten soil paramaters for a given soil type from an
internal dataset
}
\details{
This function derives the following van Genuchten paramaters used to estimate
hydrualic conductivity and diffusivity (see van Genuchten 1980 Soil Science Society of
America Journal 44:892-898). The returned list contains the following elements:
\code{n} the pore size distribution parameter (dimensionless, > 1), \code{alpha} van Genuchten
shape paramater (cm^-1), \code{Ksat} saturated hydraulic conductivity (cm / day), \code{Smin}
Residual water content (cm^3 / cm^3), \code{Smax} Water content at saturation (cm^3 / cm^3).
Each element of the list is of the same class as \code{soiltype}.
}
\examples{
getsoilparams("Loam")

}
\seealso{
\code{\link[=soilparams]{soilparams()}}
}
