% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{IMHyetosRepScheme}
\alias{IMHyetosRepScheme}
\title{Internal funcion used by \code{\link[=subdailyrain]{subdailyrain()}}}
\usage{
IMHyetosRepScheme(L, l, f, k, a, v, mx, Zhist, subtimescale)
}
\description{
Internal funcion used by \code{\link[=subdailyrain]{subdailyrain()}}
}
\keyword{internal}
