% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{IMlevel2}
\alias{IMlevel2}
\title{Internal funcion used by \code{\link[=subdailyrain]{subdailyrain()}}}
\usage{
IMlevel2(L, l, f, k, a, v, mx, Zhist, subtimescale)
}
\description{
Internal funcion used by \code{\link[=subdailyrain]{subdailyrain()}}
}
