% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydromodel.R
\name{Ktheta}
\alias{Ktheta}
\title{Hydraulic conductivity}
\usage{
Ktheta(sm, sp, timestep = 86400)
}
\arguments{
\item{sm}{Soil water fraction}

\item{sp}{a list of van Genuchten soil parameters as returned by \code{\link[=getsoilparams]{getsoilparams()}}}

\item{timestep}{number of seconds in each time step of model run (default one day)}
}
\value{
Hydraulic conductivity (cm / time-step)
}
\description{
\code{Ktheta} calculates hydrualic conductivity
}
\details{
Applies the van Genuchten equation to estimate the hydrualic conductivity of
unsaturated soil (van Genuchten 1980 Soil Science Society of America Journal 44:892-898).
Paramaters of the van Genuchten equation can be found in the dataset \code{\link[=soilparams]{soilparams()}}. Ksat
and the returned value are adjusted so that returned units are in cm per \code{timestep}.
}
\examples{
sm <- c(74:422) / 1000
plot(Ktheta(sm, getsoilparams("Loam")) ~ sm)

}
\seealso{
\code{\link[=soilparams]{soilparams()}}
}
