% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{objf1}
\alias{objf1}
\title{Internal objective funcion used to derive Bartlett-Lewis model parameters}
\usage{
objf1(x, means, vars, covs, pdrs, period = "daily")
}
\description{
Internal objective funcion used to derive Bartlett-Lewis model parameters
}
