% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{pdrMBLRPM}
\alias{pdrMBLRPM}
\title{Internal funcion used to calculate probability of being dry}
\usage{
pdrMBLRPM(a, l, v, k, f, h = 1)
}
\description{
Internal funcion used to calculate probability of being dry
}
