% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_montieth.R
\name{penman_montieth}
\alias{penman_montieth}
\title{Penman Montieth Equation}
\usage{
penman_montieth(Tair, vpd, Rnet, gs, ga, dayl, CP = 1010, Pair = 101325,
  newp = 1)
}
\arguments{
\item{Tair}{(deg C) air temperature}

\item{vpd}{(Pa)    vapour pressure deficit}

\item{Rnet}{(j/m2/day)      net radiation}

\item{gs}{(s/mm)  surface conductance}

\item{ga}{(s/mm)  aerodynamic conductance}

\item{CP}{=      1010.0 (J/kg*K) specific heat of air}

\item{Pair}{= 101325 (Pa)}

\item{newp}{sensitivity analysis default=1}
}
\value{
Evapotranspiration (mm/day)
}
\description{
THis function computer evapotranspiration based on radiation, conductance etc
}
\author{
Naomi
}

