% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{perpslope}
\alias{perpslope}
\title{Calculates slope perpendicular to direction of flow}
\usage{
perpslope(dem)
}
\arguments{
\item{dem}{a raster object of elevations.}
}
\value{
a raster object of slope angle in radians perpendicular to the direction of flow
}
\description{
\code{perpslope} is an internal function used by \code{\link[=flowvelocity]{flowvelocity()}} to calculate the slope
perpendicular to the direction of flow.
}
\examples{
dem <- aggregate(dtm1m, 20)
plot(perpslope(dem) * (180 / pi), main = "slope")

}
