% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{pressure.height}
\alias{pressure.height}
\title{Calculates atmospheric pressure from pressure at sea-level}
\usage{
pressure.height(psl = 101300, dem)
}
\arguments{
\item{psl}{an optional single numeric value, raster, two-dimensional array or matrix of
sea-level atmospheric pressures (Pa)}

\item{dem}{a single numeric value, raster, two-dimensional array or matrix of elevations (m)}
}
\value{
a single numeric value, two-dimensional array or matrix of atmospheric pressure (Pa)
}
\description{
\code{pressure.height} applies asimple height adjustment to sea-level pressure data
}
\details{
Height adjustments are performed using a simplification of the ideal gas law,
assuming 20°C for a standard atmosphere. If psl is not supplied, a typical mean value of sea-level
pressure is assumed.
}
\examples{
plot(pressure.height(dem = dtm100m))

}
