% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydromodel.R
\name{runoff}
\alias{runoff}
\title{Calculates runoff}
\usage{
runoff(rain, cn, sm, Smin, Smax)
}
\arguments{
\item{rain}{rainfall in mm}

\item{cn}{run-off curve number (see details)}

\item{sm}{Soil water fraction}

\item{Smin}{Residual water content (cm^3 / cm^3)}

\item{Smax}{Volumetric water content at saturation (cm^3 / cm^3)}
}
\value{
direct runoff in mm
}
\description{
\code{runoff} calculates direct runoff based on USDA Run-off curve number
}
\details{
Applies the United States Department of Agriculture method for
calculating direct runoff. Runoff curve numbers can be obtained from here:
https://www.wcc.nrcs.usda.gov/ftpref/wntsc/H&H/training/runoff-curve-numbers1.pdf
Adjustemnts for antecedent moisture condition are automatically applied, so the
supplied value of \code{cn} should be for average soil moisture (AMC II). \code{Smin} and \code{Smax}
values for soil types are given in \code{\link[=soilparams]{soilparams()}}.
}
\examples{
rain <- c(1:150)
ro1 <- runoff(rain, cn = 50, sm = 0.4, Smin = 0.074, Smax = 0.422)
ro2 <- runoff(rain, cn = 90, sm = 0.4, Smin = 0.074, Smax = 0.422)
plot(ro1 ~ rain, type = "l")
plot(ro2 ~ rain, type = "l")
}
