% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydromodel.R
\name{soilevap}
\alias{soilevap}
\title{Soil evapotranspiration}
\usage{
soilevap(cre, sm, cover = 0.5, topratio = 0.5)
}
\arguments{
\item{cre}{Crop reference evapotranspiration as returned by \code{\link[=cre.hourly]{cre.hourly()}} or \code{\link[=cre.daily]{cre.daily()}}}

\item{sm}{Soil water fraction}

\item{cover}{Fraction of vegetation cover}

\item{topratio}{Ratio of root water uptake from top layer relative to bottom layer
(1 = all from top layer, 0 = all form bottom layer)}
}
\value{
a list of with two elements: \code{top} evaporation and evapotranspiration from
the top soil layer, \code{btm} evapotranspiration from the bottom layer. Units are as for \code{CRE}
}
\description{
Soil evapotranspiration
}
\details{
\code{soilevap} calculates soil evapotranspiration from crop reference evapotranspiration.

As the soil water fraction decreases below the saturation value, evaporation from
bare soil is usually assumed to continue at the potential rate until the water content
decreases below a critical value. Thereafter, evaporation is assumed to decrease linearly
with decreasing water content, vanishing to zero when water content is low. The paramater
\code{cover} aportions the ratio of evapotranspiration to bare soil evaporation. The paramater
\code{topratio} aportions evapotranspiration between the top and bottom layer and depends on
the chosen soil layer depths and root depths.
}
\examples{
sm <- c(0:100) / 100
plot(soilevap(1, sm, cover = 1, surface = 1)$top ~ sm, type = "l")
plot(soilevap(1, sm, cover = 0, surface = 1)$top ~ sm, type = "l")
}
\seealso{
\code{\link[=cre.hourly]{cre.hourly()}} \code{\link[=cre.daily]{cre.daily()}}
}
