% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{topdist}
\alias{topdist}
\title{Distributes soil moisture by topographic wetness}
\usage{
topdist(tx, sm, p = 0.25, Smin = 0.091, Smax = 0.419)
}
\arguments{
\item{tx}{a vector or marix of topographic wetness index values}

\item{sm}{a single numeric value of the mean soil water fraction}

\item{p}{a single numeric value of to power adjustment to apply to topographic wetness
index values (see details)}

\item{Smin}{a single numeric value of the residual soil water fraction (see \code{\link[=soilparams]{soilparams()}})}

\item{Smax}{a single numeric value of the saturated  soil water fraction (see \code{\link[=soilparams]{soilparams()}})}
}
\value{
a vector or matrix of soil moisture values of each grid cell of tx
}
\description{
\code{topdist} distributes soil moisture, typically within a basin, by the topograhic wetness index of each basin
grid cell
}
\details{
the coefficient p scales the relationship between topographic wetness and the
distribution of soil moisture values. If p = 1, soil moisture is strongly concentrated in
cells with a high topographic wetness. If p = 0, soil moisture is evenly distributed across
the basin. The correct choice of p depends on sub-grid scale variation in topographic wetness, and on
complex lateral flows, and thus best determined empirically, by measuring soil moisture across
a catchment. A choice of 0.25 for surface soil layers and 0.15 for sub-surface soil layers
is likely to provide a reasonable approximation.
}
\examples{
library(raster)
tpx <- topidx(dtm100m)
tp1 <- topdist(is_raster(tpx), 0.3, p = 0.25)
tp2 <- topdist(is_raster(tpx), 0.3, p = 0.15)
# NB - not distributed by basin
plot(if_raster(tp1, dtm100m))
plot(if_raster(tp2, dtm100m))
}
\seealso{
\code{\link[=topidx]{topidx()}}, \code{\link[=soilparams]{soilparams()}}
}
