% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{topdistw}
\alias{topdistw}
\title{Distributes surface water by topographic wetness}
\usage{
topdistw(tx, wvol, p = 0.3, xres, yres)
}
\arguments{
\item{tx}{a vector or matrix of topographic wetness index values}

\item{wvol}{a single numeric value of basin  water volume (m^3)}

\item{p}{a single numeric value of to power adjustment to apply to topographic wetness
index values (see details)}

\item{xres}{x grid cell resolution (m)}

\item{yres}{y grid cell resolution (m)}
}
\value{
a vector or matrix of surface water depth (mm) for each grid cell of tx
}
\description{
\code{topdistw} surface water, typically within a basin, by the topograhic wetness index of each basin
grid cell
}
\details{
the coefficient p scales the relationship between topographic wetness and the
distribution of surface water. If p = 1, soil moisture is strongly concentrated in
cells with a high topographic wetness. If p = 0, soil moisture is evenly distributed across
the basin. The correct choice of p depends on sub-grid scale variation in basin depressions, and on
complex lateral flows, and thus best determined empirically. A choice of 0.3 for is likely to provide a reasonable approximation.
}
\examples{
library(raster)
tpx <- topidx(dtm100m)
# average depth on 10 mm
sel <- which(is.na(is_raster(tpx) == F))
wvol <- 10 * 100 * 100 * length(sel) / 1000
# NB - not distributed by basin
swd1 <- topdistw(is_raster(tpx), wvol, p = 0.3, xres = 100, yres = 100)
swd2 <- topdistw(is_raster(tpx), wvol, p = 0.1, xres = 100, yres = 100)
plot(if_raster(swd1, dtm100m))
plot(if_raster(swd2, dtm100m))
}
\seealso{
\code{\link[=topidx]{topidx()}}
}
