% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basintools.R
\name{topidx}
\alias{topidx}
\title{Calculates Bevan and Kirkby's topographic wetness index}
\usage{
topidx(dem, minslope = atan(0.005/mean(res(dem))))
}
\arguments{
\item{dem}{a raster object of elevations.}

\item{minslope}{an optional positive value specifying the minimum slope allowed (see details).}
}
\value{
a raster object of topographic wetness.
}
\description{
\code{topidx} is used to calculate the topographic wetness index described in Bevan and Kirkby (1979)
basin
}
\details{
If slope is zero, infinite topographic wetness index values are returned. Thus,
flat areas are assigned a value specifed by \code{minslope}. The default value, given by
atan(0.005 / mean(res(dem))), assumes that height differences between adjacent cells exceed
0.005. Accumulated flow is multiplied by the cell dimensions, thus\code{dem} should have an
equal area projection.
}
\examples{
dem <- aggregate(dtm1m, 20)
plot(log(topidx(dem)), main = "Topographic wetness")

}
\references{
Bevan KJ & Kirkby MJ (1979) A physically based, variable contributing area
model of basin hydrology. Hydrological Sciences Journal 24: 43-69.
}
