% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mettools.R
\name{varMBLRPM}
\alias{varMBLRPM}
\title{Internal funcion used to calculate variance of rainfall}
\usage{
varMBLRPM(a, l, v, k, f, mx, h = 1)
}
\description{
Internal funcion used to calculate variance of rainfall
}
