% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruit.R
\name{compute_yield}
\alias{compute_yield}
\title{compute annual yield'}
\usage{
compute_yield(T, P, ts, tp, intercept, irr = "N")
}
\arguments{
\item{T}{annual temperature (C)}

\item{P}{annual precipitation (mm)}

\item{ts}{slope on temperature}

\item{tp}{slope on precipitation}

\item{intercept}{(kg)}

\item{irr}{Y or  N (default N)}
}
\value{
yield in kg
}
\description{
Function to compute yeild of different fruits as a function of annual temperature and precipitation
}

